Refactor(#151): class Info 받는 중복 코드 수정
