+++ src/entities/Class/ClassInfo/index.tsx
@@ -33,6 +33,7 @@ export const ClassInfo: React.FC<ClassInfoProps> = ({
       setIsLoading(true);
       
       const response = await getClassInfo(classId);
+      console.log("넘어온 값",response)
       
       if (response && typeof response === 'object') {
         setClassData(response);
