+++ src/pages/Teacher/Class/index.tsx
@@ -7,7 +7,8 @@ import TabSelector from '@/entities/UI/TabSelect';
 import LessonComponent from '@/features/Common/Class/Lesson';
 import { AssignmentComponent } from '@/features/Common/Class/Assignment';
 import { ExamComponent } from '@/features/Common/Class/Exam';
-import { ClassData, tabs } from './data';
+import { tabs } from './data';
+import { ClassData } from '@/features/Common/Main/hooks/useClass';
 import { ClassInfoProps } from '@/shared/types/Class/classroom';
 import NotFound from '@/pages/NotFound';
 
@@ -17,6 +18,7 @@ const Classroom: React.FC = () => {
 
   // API로 가져온 클래스 정보를 담을 state
   const [classInfo, setClassInfo] = useState<ClassInfoProps | null>(null);
+  console.log(classInfo);
 
   useEffect(() => {
     const savedTab = localStorage.getItem('classroom-tab');
@@ -50,7 +52,7 @@ const Classroom: React.FC = () => {
   const renderContent = () => {
     switch (activeTab) {
       case 'lesson':
-        return <LessonComponent classId={classRoomId} />;
+        return <LessonComponent classRoomId={classRoomId} />;
       case 'assignment':
         return <AssignmentComponent />;
       case 'exam':
